# proto/gen/main.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "main.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Calculator"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/calculator/add": {
      "post": {
        "summary": "Simple addition over HTTP POST",
        "operationId": "Calculator_Add",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/calculatorAddResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/calculatorAddRequest"
            }
          }
        ],
        "tags": [
          "Calculator"
        ]
      }
    },
    "/v1/calculator/fibonacci": {
      "get": {
        "summary": "Fibonacci over HTTP GET with query param",
        "operationId": "Calculator_GenerateFibonacci",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/calculatorFibonacciResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of calculatorFibonacciResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "n",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Calculator"
        ]
      }
    }
  },
  "definitions": {
    "calculatorAddRequest": {
      "type": "object",
      "properties": {
        "a": {
          "type": "integer",
          "format": "int32",
          "title": "must be \u003e 0"
        },
        "b": {
          "type": "integer",
          "format": "int32",
          "title": "must be \u003e 0"
        }
      }
    },
    "calculatorAddResponse": {
      "type": "object",
      "properties": {
        "sum": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "calculatorChatMessage": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "calculatorFibonacciResponse": {
      "type": "object",
      "properties": {
        "number": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "calculatorNumberResponse": {
      "type": "object",
      "properties": {
        "sum": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}